requirements.txt
Django==4.2.1
djangorestframework~=3.14.0
psycopg2
requests~=2.31.0